Ensure we do not evict txs before `last_valid`
When updating a `ChainGraph`, we evict txs of the original chain of it
conflicts with any of the update txs. However, evicting transactions
before `last_valid` checkpoint means that we have inconsistent
histories. Thus, we shoud return an error.

Additionaly, `SparseChain::determine_changeset` returns a tuple
`(changeset, last_valid)` so that `last_valid` could be used for the
aforementioned checks.