Test `ChainGraph`'s eviction logic
As mentioned by @LLFourn in #76:

I think #74 also introduced a bug where an update can evict a
transaction in the chain without invalidating its block by having a new
transaction that conflicts with it. We shouldn't allow this and this
should mean the update is invalid. Test for this too.

Additionally, `ChainGraph` logic now checks both graphs (original and
update) for full tx during determining txs to evict.