Improve file store ergonomics
We need to be able to open a file without loading it yet.

Added is_empty methods on some types that I didn't end up using but kept
them because it seemed helpful.

Removed the caching of derivation indices within the store because it
wasn't really doing it's job with the cli example and we should do this
properly on the outside.

Removed the Error for ChainGraph::apply_changeset because if we're not
guaranteeing everything we shouldn't try and guarantee anything wrt to
changeset.