Add type parameter to TxGraph<T>
This allows create a TxGraph<&Transaction> or a
TxGraph<Cow<Transaction>> etc. It also allows application specific
transaction types.

The main motivation for this is allowing a ChainGraph<_
Cow<Transaction>> so we can inflate a sparse chain to a chain graph from
another chain graph without copying transactions between them.