Fix `KeychainTxOutIndex` lookahead logic
The lookahead logic tries to be as efficient as possible and avoids
storing spks that are already stored. However, the original logic failed
when the lookahead setting was changed more than once. This fixes it.

Additionally, the methods `lookahead_to_target` and
`lookahead_to_target_multi` are added. These store unrevealed scripts
without affecting lookahead settings.